Add bodyCacheFile.
If `parseReqBody` is `false`, there is was no way to get the content of the body.
If you read from the request in `decorateRequest`, the request stream would be
empty when sending the upstream the request.
If you read from the request in `intercept`, the request stream would be
empty as it was already drained sending the upstream request.
There was no way to tap into the request -> upstream pipe.

This adds an option to write the body to a file, making the contents of
the request available to `decorateRequest` and `intercept`.